Step back with .gitignore problem
Tried to access my_facebook_access.py directly in the folder just by
renaming the example model but without uploading on the repository as
it needs to remain private because of our Facebook private information.
Unfortunately, I couldn’t guarantee that the file wouldn’t upload and,
to make sure, took a step back, returned to the “path to Desktop” way
to import the file my_facebook_access.py.